+++ client/src/components/Cards/trees.tsx
@@ -1,8 +1,9 @@
-import { useState } from "react";
 import { Tree } from "../../types/tree";
 import { CardContainer, CardContent, CardHeader, CardOptions } from "./styles";
 import { PiTreeDuotone } from "react-icons/pi";
 import { api } from "../../services/api";
+import { useContext } from "react";
+import { ModalContext } from "../../contexts/ModalContext";
 
 interface TreeCardProps {
     getDeleted: (deleted: string) => void;
@@ -25,21 +26,47 @@ const TreeCard = (props: TreeCardProps) => {
         species
     } = tree;
 
-    const [deleted, setDeleted] = useState<string>('');
+    const { setShow, setType, setMessage } = useContext(ModalContext);
 
     const deleteTree = async (id: string) => {
-        api.delete('/trees/delete/deleteTree', {
-            params: {
-                id: id
-            }
-        })
-        .then(() => {
-            setDeleted(id);
-            getDeleted(deleted);
-        })
-        .catch((err) => {
-            console.error(err);
-        });
+        if (window.confirm('Colheitas talvez sejam excluídas.\nDeseja prosseguir?')) {
+            api.delete('/trees/delete/deleteTree', {
+                params: {
+                    id: id
+                }
+            })
+            .then(() => {
+                getDeleted(id);
+                setShow(true);
+                setType('success');
+                setMessage('Árvore excluída.');
+            })
+            .catch((err) => {
+                console.error(err);
+                setShow(true);
+                setType('error');
+                setMessage(err.response.data);
+            });
+        }
+    }
+
+    const formatAge = (age: number) => {
+        const years = Math.floor(age / 12);
+        const months = age % 12;
+        if (years > 0 && months > 0) {  // mês e ano
+            return (
+                `${years === 1 ? years + ' ano e ' : years + ' anos e '}` // singular e plural
+                +
+                `${months === 1 ? months + ' mês' : months + ' meses'}`   // singular e plural
+            );
+        }
+        if (years > 0) {    // apenas ano
+            return `${years === 1 ? years + ' ano' : years + ' anos'}`;
+        }
+        if (months > 0) {   // apenas mês
+            return `${months === 1 ? months + ' mês' : months + ' meses'}`;
+        }
+        return '0'; 
     }
 
     return (
@@ -48,18 +75,22 @@ const TreeCard = (props: TreeCardProps) => {
                 <p>Árvore <PiTreeDuotone /></p>
                 <p>#{id.slice(0, 8)}</p>
             </CardHeader>
+
             <CardContent>
                 <p>Descrição:</p>
                 <p>{description}</p>
             </CardContent>
+
             <CardContent>
                 <p>Idade:</p>
-                <p>{age}</p>
+                <p>{formatAge(age)}</p>
             </CardContent>
+
             <CardContent>
                 <p>Espécie:</p>
                 <p>{species.description}</p>
             </CardContent>
+            
             <CardOptions>
                 <p className="edit" onClick={() => getToEdit(tree)}>Editar</p>
                 <p className="delete" onClick={() => deleteTree(id)}>Excluir</p>
